+++ src/v/cloud_topics/BUILD
@@ -6,11 +6,11 @@ package(default_visibility = [
     "//src/v/cloud_topics/core/tests:__pkg__",
     "//src/v/cloud_topics/frontend:__pkg__",
     "//src/v/cloud_topics/level_zero:__subpackages__",
+    "//src/v/cloud_topics/level_zero/throttler:__pkg__",
+    "//src/v/cloud_topics/level_zero/throttler/tests:__pkg__",
     "//src/v/cloud_topics/read_path:__pkg__",
     "//src/v/cloud_topics/read_path/tests:__pkg__",
     "//src/v/cloud_topics/tests:__pkg__",
-    "//src/v/cloud_topics/throttler:__pkg__",
-    "//src/v/cloud_topics/throttler/tests:__pkg__",
     "//src/v/kafka/data:__pkg__",
 ])
 
@@ -154,12 +154,12 @@ redpanda_cc_library(
         "//src/v/cloud_topics:data_plane_api",
         "//src/v/cloud_topics:ephemeral_cluster_services",
         "//src/v/cloud_topics/batch_cache",
-        "//src/v/cloud_topics/core:read_pipeline",
-        "//src/v/cloud_topics/core:write_pipeline",
+        "//src/v/cloud_topics/level_zero:read_pipeline",
+        "//src/v/cloud_topics/level_zero:write_pipeline",
         "//src/v/cloud_topics/level_zero/batcher",
         "//src/v/cloud_topics/level_zero/reconciler",
+        "//src/v/cloud_topics/level_zero/throttler",
         "//src/v/cloud_topics/read_path:fetch_handler",
-        "//src/v/cloud_topics/throttler",
         "//src/v/model",
         "//src/v/storage",
     ],
