--- src/v/cloud_topics/core/README.md
@@ -1,8 +1,0 @@
-The `core` contains utilities shared by read path, write path and resource balancer.
-It includes the following components:
-- `read_request`
-- `write_request`
-- `serializer`
-- `write_pipeline`
-- `read_pipeline`
-- etc
