+++ src/v/cloud_topics/level_zero/batcher/BUILD
@@ -16,7 +16,7 @@ redpanda_cc_library(
     implementation_deps = [
         "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:logger",
-        "//src/v/cloud_topics/core:serializer",
+        "//src/v/cloud_topics/level_zero:serializer",
         "//src/v/container:chunked_circular_buffer",
         "//src/v/storage:record_batch_builder",
     ],
@@ -25,7 +25,7 @@ redpanda_cc_library(
         "//src/v/base",
         "//src/v/cloud_topics:cluster_services_interface",
         "//src/v/cloud_topics:types",
-        "//src/v/cloud_topics/core:write_request",
+        "//src/v/cloud_topics/level_zero:write_request",
         "//src/v/container:fragmented_vector",
         "//src/v/model",
         "@abseil-cpp//absl/container:btree",
@@ -47,7 +47,7 @@ redpanda_cc_library(
         "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:logger",
         "//src/v/cloud_topics:object_utils",
-        "//src/v/cloud_topics/core:serializer",
+        "//src/v/cloud_topics/level_zero:serializer",
         "//src/v/ssx:sformat",
         "//src/v/utils:human",
     ],
@@ -58,10 +58,10 @@ redpanda_cc_library(
         "//src/v/bytes:iobuf",
         "//src/v/cloud_topics:cluster_services_interface",
         "//src/v/cloud_topics:types",
-        "//src/v/cloud_topics/core:event_filter",
-        "//src/v/cloud_topics/core:pipeline_stage",
-        "//src/v/cloud_topics/core:write_pipeline",
-        "//src/v/cloud_topics/core:write_request",
+        "//src/v/cloud_topics/level_zero:event_filter",
+        "//src/v/cloud_topics/level_zero:pipeline_stage",
+        "//src/v/cloud_topics/level_zero:write_pipeline",
+        "//src/v/cloud_topics/level_zero:write_request",
         "//src/v/config",
         "//src/v/model",
         "//src/v/utils:retry_chain_node",
