+++ src/v/cloud_topics/level_zero/batcher/batcher.cc
@@ -12,11 +12,11 @@
 
 #include "base/unreachable.h"
 #include "cloud_io/remote.h"
-#include "cloud_topics/core/event_filter.h"
-#include "cloud_topics/core/serializer.h"
-#include "cloud_topics/core/write_request.h"
 #include "cloud_topics/errc.h"
 #include "cloud_topics/level_zero/batcher/aggregator.h"
+#include "cloud_topics/level_zero/event_filter.h"
+#include "cloud_topics/level_zero/serializer.h"
+#include "cloud_topics/level_zero/write_request.h"
 #include "cloud_topics/logger.h"
 #include "cloud_topics/object_utils.h"
 #include "cloud_topics/types.h"
