+++ src/v/cloud_topics/level_zero/batcher/batcher.h
@@ -16,8 +16,8 @@
 #include "base/units.h"
 #include "bytes/iobuf.h"
 #include "cloud_topics/cluster_services.h"
-#include "cloud_topics/core/pipeline_stage.h"
-#include "cloud_topics/core/write_pipeline.h"
+#include "cloud_topics/level_zero/pipeline_stage.h"
+#include "cloud_topics/level_zero/write_pipeline.h"
 #include "cloud_topics/types.h"
 #include "config/property.h"
 #include "model/fundamental.h"
