+++ src/v/cloud_topics/level_zero/circuit_breaker.h
