+++ src/v/cloud_topics/level_zero/pipeline_stage.h
