+++ src/v/cloud_topics/level_zero/read_request.cc
@@ -8,9 +8,9 @@
  * https://github.com/redpanda-data/redpanda/blob/master/licenses/rcl.md
  */
 
-#include "cloud_topics/core/read_request.h"
+#include "cloud_topics/level_zero/read_request.h"
 
-#include "cloud_topics/core/pipeline_stage.h"
+#include "cloud_topics/level_zero/pipeline_stage.h"
 #include "cloud_topics/logger.h"
 #include "utils/retry_chain_node.h"
