+++ src/v/cloud_topics/level_zero/read_request.h
@@ -12,9 +12,9 @@
 
 #include "base/outcome.h"
 #include "base/seastarx.h"
-#include "cloud_topics/core/pipeline_stage.h"
 #include "cloud_topics/errc.h"
 #include "cloud_topics/extent_meta.h"
+#include "cloud_topics/level_zero/pipeline_stage.h"
 #include "container/fragmented_vector.h"
 #include "container/intrusive_list_helpers.h"
 #include "model/record.h"
