+++ src/v/cloud_topics/level_zero/serializer.h
