+++ src/v/cloud_topics/level_zero/tests/write_pipeline_test.cc
@@ -8,8 +8,8 @@
  * https://github.com/redpanda-data/redpanda/blob/master/licenses/rcl.md
  */
 
-#include "cloud_topics/core/write_pipeline.h"
 #include "cloud_topics/extent_meta.h"
+#include "cloud_topics/level_zero/write_pipeline.h"
 #include "container/chunked_circular_buffer.h"
 #include "model/fundamental.h"
 #include "model/namespace.h"
