+++ src/v/cloud_topics/level_zero/throttler/tests/BUILD
@@ -9,10 +9,10 @@ redpanda_cc_gtest(
     deps = [
         "//src/v/base",
         "//src/v/cloud_topics:extent_meta",
-        "//src/v/cloud_topics/core:pipeline_stage",
-        "//src/v/cloud_topics/core:write_pipeline",
-        "//src/v/cloud_topics/core:write_request",
-        "//src/v/cloud_topics/throttler",
+        "//src/v/cloud_topics/level_zero:pipeline_stage",
+        "//src/v/cloud_topics/level_zero:write_pipeline",
+        "//src/v/cloud_topics/level_zero:write_request",
+        "//src/v/cloud_topics/level_zero/throttler",
         "//src/v/container:chunked_circular_buffer",
         "//src/v/model",
         "//src/v/model/tests:random",
