+++ src/v/cloud_topics/level_zero/throttler/throttler.cc
@@ -8,12 +8,12 @@
  * https://github.com/redpanda-data/redpanda/blob/master/licenses/rcl.md
  */
 
-#include "cloud_topics/throttler/throttler.h"
+#include "cloud_topics/level_zero/throttler/throttler.h"
 
 #include "base/unreachable.h"
-#include "cloud_topics/core/event_filter.h"
-#include "cloud_topics/core/write_pipeline.h"
-#include "cloud_topics/core/write_request.h"
+#include "cloud_topics/level_zero/event_filter.h"
+#include "cloud_topics/level_zero/write_pipeline.h"
+#include "cloud_topics/level_zero/write_request.h"
 #include "cloud_topics/logger.h"
 
 #include <seastar/core/loop.hh>
