+++ src/v/cloud_topics/level_zero/write_request.h
@@ -12,10 +12,10 @@
 
 #include "base/outcome.h"
 #include "base/seastarx.h"
-#include "cloud_topics/core/pipeline_stage.h"
-#include "cloud_topics/core/serializer.h"
 #include "cloud_topics/errc.h"
 #include "cloud_topics/extent_meta.h"
+#include "cloud_topics/level_zero/pipeline_stage.h"
+#include "cloud_topics/level_zero/serializer.h"
 #include "container/fragmented_vector.h"
 #include "container/intrusive_list_helpers.h"
