+++ src/v/cloud_topics/read_path/BUILD
@@ -87,9 +87,9 @@ redpanda_cc_library(
         "//src/v/cloud_topics:logger",
         "//src/v/cloud_topics:placeholder",
         "//src/v/cloud_topics:types",
-        "//src/v/cloud_topics/core:event_filter",
-        "//src/v/cloud_topics/core:read_pipeline",
-        "//src/v/cloud_topics/core:read_request",
+        "//src/v/cloud_topics/level_zero:event_filter",
+        "//src/v/cloud_topics/level_zero:read_pipeline",
+        "//src/v/cloud_topics/level_zero:read_request",
         "//src/v/cloud_topics/read_path:materialized_extent_reader",
         "//src/v/model",
         "//src/v/ssx:future_util",
