+++ src/v/cloud_topics/read_path/tests/BUILD
@@ -91,7 +91,7 @@ redpanda_cc_gtest(
     deps = [
         "//src/v/base",
         "//src/v/cloud_topics:types",
-        "//src/v/cloud_topics/core:read_pipeline",
+        "//src/v/cloud_topics/level_zero:read_pipeline",
         "//src/v/cloud_topics/read_path:fetch_handler",
         "//src/v/cloud_topics/read_path/tests:materialized_extent_fixture",
         "//src/v/container:fragmented_vector",
