+++ src/v/cloud_topics/read_path/tests/fetch_handler_test.cc
@@ -8,8 +8,8 @@
  * https://github.com/redpanda-data/redpanda/blob/master/licenses/rcl.md
  */
 
-#include "cloud_topics/core/read_pipeline.h"
 #include "cloud_topics/errc.h"
+#include "cloud_topics/level_zero/read_pipeline.h"
 #include "cloud_topics/read_path/fetch_request_handler.h"
 #include "cloud_topics/read_path/tests/materialized_extent_fixture.h"
 #include "container/fragmented_vector.h"
